python_microservices/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
cloudinary
Pillow
google-generativeai
requests
bson
python-multipart
serpapi
praw